src/main/java/com/example/memberweb/service/MemberService.java
package com.example.memberweb.service;

import com.example.memberweb.domain.Member;
import com.example.memberweb.repository.MemberRepository;
import com.example.memberweb.repository.MemoryMemberRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;
import java.util.List;

//@Service
public class MemberService {

    private final MemberRepository memberRepository;

    //MemberRepository에 연결시켜준다.
    //@Autowired
    public MemberService(MemberRepository memberRepository){
        this.memberRepository = memberRepository;
    }

    //회원 가입
    public Long join(Member member){
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    //중복 확인
    private void validateDuplicateMember(Member member) {
        Optional<Member> result = memberRepository.findByName(member.getName());
        result.ifPresent(m -> {
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        });
    }

    //전체 회원 조회
    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

    //회원 조회
    public Optional<Member> findMember(Long memberId){
        return memberRepository.findById(memberId);
    }


}
